feat(ta): Server Component shell pre-fetches initial data to eliminate loading skeleton flash (#693)
- Add `initialData?: unknown` option to `usePolling` so Server Component
  pre-fetched data seeds the hook's initial state.  Cache takes precedence
  when warm (returning visitor), otherwise `initialData` prevents the
  null → data transition that causes the loading skeleton flash.
- Create `src/lib/ta/initial-data.ts`: `fetchTaInitialData(id)` runs the
  same Prisma queries as GET /api/tournaments/[id]/ta plus the players
  list in a single parallel batch; returns null on error so the client
  falls back gracefully to its own first poll.
- Rename `ta/page.tsx` → `ta/page-client.tsx`; component now accepts
  `tournamentId: string` and `initialData` props instead of `use(params)`.
- New `ta/page.tsx` is a `force-dynamic` Server Component that awaits
  params, calls `fetchTaInitialData`, and renders the client shell with
  `initialData`.
- 4 new unit tests for the `initialData` option (initial value, cache
  wins over server data, null fallback, first poll replaces server data).
